+++ app/src/main/java/org/intelehealth/ezazi/models/ActivePatientModel.java
@@ -17,7 +17,7 @@ public class ActivePatientModel {
     String phone_number;
     String sync;
     String gender;
-    int alertFlagTotal;
+    double alertFlagTotal;
     int visibilityOrder;
     private String stageName;
     private boolean obsExistsFlag;
@@ -141,11 +141,11 @@ public class ActivePatientModel {
         this.gender = gender;
     }
 
-    public int getAlertFlagTotal() {
+    public double getAlertFlagTotal() {
         return alertFlagTotal;
     }
 
-    public void setAlertFlagTotal(int alertFlagTotal) {
+    public void setAlertFlagTotal(double alertFlagTotal) {
         this.alertFlagTotal = alertFlagTotal;
     }
